add server.go

+++ model/customer.go
@@ -2,5 +2,5 @@ package model
 
 type Customer struct {
 	LopeiId int32
-	Balance int32
+	Balance float32
 }
